requirements.txt
# Function dependencies, for example:
# package>=version

web3
python-dotenv
